+++ src/app.module.ts
@@ -1,19 +1,19 @@
 import { Module } from '@nestjs/common';
 import { ConfigService } from '@nestjs/config';
 import { APP_FILTER, APP_GUARD, APP_INTERCEPTOR } from '@nestjs/core';
-import { AppController } from './app.controller';
-import { AppService } from './app.service';
-import { SuccessInterceptor } from './common/success-response.interceptor';
-import { PrismaModule } from './database/prisma/prisma.module';
-import { GlobalExceptionFilter } from './exceptions/global-exception.filter';
-import { AuthGuard } from './guard/auth.guard';
-import { UsersController } from './users/users.controller';
-import { UsersModule } from './users/users.module';
-import { UsersService } from './users/users.service';
-import { S3Service } from './users/storage/aws.service';
-import { NestjsFormDataModule } from 'nestjs-form-data';
 import { MulterModule } from '@nestjs/platform-express';
 import * as multer from 'multer';
+import { NestjsFormDataModule } from 'nestjs-form-data';
+import { AppController } from './app.controller';
+import { AppService } from './app.service';
+import { GlobalExceptionFilter } from './common/filters/global-exception.filter';
+import { SuccessInterceptor } from './common/interceptors/success-response.interceptor';
+import { PrismaModule } from './database/prisma/prisma.module';
+import { S3Service } from './domains/users/storage/aws.service';
+import { UsersController } from './domains/users/users.controller';
+import { UsersModule } from './domains/users/users.module';
+import { UsersService } from './domains/users/users.service';
+import { AuthGuard } from './guard/auth.guard';
 
 @Module({
   imports: [
