+++ src/common/interceptors/profile-image-upload.interceptor.ts
