+++ src/common/interceptors/success-response.interceptor.ts
