+++ src/decorator/private.decorator.ts
@@ -1,5 +1,5 @@
 import { SetMetadata } from '@nestjs/common';
-import { AccountType } from 'src/users/users.type';
+import { AccountType } from 'src/domains/users/users.type';
 
 export const Private = (accountType: AccountType) =>
   SetMetadata('accountType', accountType);
