+++ src/domains/users/users.module.ts
@@ -5,7 +5,7 @@ import { UsersService } from './users.service';
 import { S3Module } from './storage/aws.module';
 import { NestjsFormDataModule } from 'nestjs-form-data';
 import { S3Service } from './storage/aws.service';
-import { ProfileImageUploadInterceptor } from 'src/interceptors/profile-image-upload.interceptor';
+import { ProfileImageUploadInterceptor } from 'src/common/interceptors/profile-image-upload.interceptor';
 
 @Module({
   controllers: [UsersController],
